feat: enable PPR and add warmup health check endpoint
- Enable Next.js cacheComponents (PPR) in next.config.ts
- Refactor stock-items page to Server Component with Suspense + skeleton
- Add /api/health route with Go backend connectivity check (5s timeout)
- Add GitHub Actions keep-warm job (every 5 min via VERCEL_APP_URL secret)

Closes #87
