+++ openspec/changes/ppr-and-warmup-health/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-05-21
