+++ openspec/changes/ppr-and-warmup-health/design.md
@@ -1,0 +1,66 @@
+## Context
+
+Vercel にデプロイした Next.js フロントエンドは散発的なアクセスパターン（家族向け）のため、アイドル後のコールドスタートが体感速度に影響している。Fluid Compute は既に有効。バックエンドは Go (AWS Lambda) で別途コールドスタートの問題がある。
+
+## Goals / Non-Goals
+
+**Goals:**
+- Next.js PPR により静的シェルを CDN から即時配信し、コールドスタートの体感影響をゼロに近づける
+- `/api/health` でフロントエンド関数と Go バックエンドの両方の死活確認を一元化する
+- GitHub Actions スケジュールジョブで Vercel 関数をウォームに維持する
+
+**Non-Goals:**
+- Go バックエンド (Lambda) のウォームアップは対象外（別課題）
+- E2E 監視・アラートの整備は対象外
+- Vercel Pro へのアップグレードは前提にしない（Hobby プランで完結）
+
+## Decisions
+
+### D1: ウォームアップ手段として GitHub Actions を選択
+
+**選択肢:**
+- A: Vercel Cron Jobs → Hobby プランでは 1日1回が上限。ウォームアップ目的には不十分
+- B: cron-job.org などの外部サービス → 無料で高頻度だが、管理場所がリポジトリ外に分散する
+- C: **GitHub Actions スケジュールジョブ** → リポジトリ内で完結、CI/CD と同じ場所で管理できる
+
+**決定:** C を採用。private リポジトリでも 2,000 分/月の無料枠があり、10分間隔（月 ~4,300 回 × ~10秒）なら ~700 分程度に収まる。
+
+### D2: `/api/health` でバックエンド疎通確認を含める
+
+**選択肢:**
+- A: フロントエンドの稼働確認のみ（`{ status: "ok" }` を返す）
+- B: **バックエンドへの HTTP 疎通確認を含める** → バックエンドの死活も同時にウォームアップできる
+
+**決定:** B を採用。Go Lambda もコールドスタートが発生するため、ヘルスチェックで同時にウォームアップする効果がある。タイムアウトは 5 秒に設定し、バックエンド障害時は degraded 状態として 200 を返す（ウォームアップ自体は成功扱い）。
+
+### D3: PPR の Suspense 境界は在庫リスト部分のみに絞る
+
+**選択肢:**
+- A: ページ全体を動的にする（現状維持）
+- B: **ヘッダー・ナビを静的、在庫リストのみを動的（Suspense）にする**
+- C: ISR で全体をキャッシュする → リアルタイム性が失われるため不採用
+
+**決定:** B を採用。ユーザーが最初に目にするレイアウトを即時表示し、データフェッチ待ちはスケルトン UI で体験を補う。
+
+## Risks / Trade-offs
+
+| リスク | 緩和策 |
+|--------|--------|
+| GitHub Actions のスケジュール遅延（数分〜数十分ずれる可能性） | ウォームアップはベストエフォート。遅延しても致命的ではない |
+| バックエンドがダウン中でもヘルスチェックが 200 を返す | レスポンスの `backend.status` フィールドで区別可能。アラートが必要になったら対応 |
+| PPR の Suspense 境界追加で既存レンダリングに影響が出る可能性 | 変更対象を在庫リストコンポーネントに限定し、影響範囲を最小化する |
+
+## Migration Plan
+
+1. `next.config.ts` に `experimental.ppr: true` を追加
+2. 在庫リストコンポーネントを `<Suspense>` で囲い、スケルトン UI を追加
+3. `app/api/health/route.ts` を新規作成
+4. `.github/workflows/keep-warm.yml` を新規作成
+5. GitHub Secrets に `VERCEL_APP_URL` を登録（手動、ユーザー作業）
+6. Vercel にデプロイして動作確認
+
+ロールバック: `next.config.ts` の `ppr` フラグを削除、ワークフローファイルを削除するだけで元に戻せる。
+
+## Open Questions
+
+- Go バックエンドの `/health` エンドポイントは既に存在するか？存在しない場合は別途追加が必要
